Colocate the genuid fn to its only caller
The create route for the tool
